+++ lib/in_app_navigator.dart
@@ -1,4 +1,6 @@
 library in_app_navigator;
 
 export 'src/delegate.dart';
+export 'src/extension.dart';
+export 'src/flags.dart';
 export 'src/router.dart';