+++ lib/src/flags.dart
@@ -1,0 +1,6 @@
+enum Flag {
+  none,
+  clear,
+  neglect,
+  replace,
+}