+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1
+
+* BuildContext to navigate supported
+* Clear all background routes and open new route
+* Easiest implementation
+
 ## 1.0.0
 
 * initial release.